moved unit-test of execute options into package options
