+++ internal/query/transaction_test.go
@@ -19,16 +19,11 @@ import (
 	grpcCodes "google.golang.org/grpc/codes"
 	grpcStatus "google.golang.org/grpc/status"
 
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/allocator"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/params"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/options"
-	internal "github.com/ydb-platform/ydb-go-sdk/v3/internal/query/tx"
 	baseTx "github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xtest"
 	"github.com/ydb-platform/ydb-go-sdk/v3/query"
-	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
-	"github.com/ydb-platform/ydb-go-sdk/v3/table/stats"
 )
 
 var _ baseTx.Transaction = &Transaction{}
@@ -494,193 +489,4 @@ func TestRollback(t *testing.T) {
 		require.Error(t, err)
 		require.True(t, xerrors.IsOperationError(err, Ydb.StatusIds_UNAVAILABLE))
 	})
-}
-
-type testExecuteSettings struct {
-	execMode    options.ExecMode
-	statsMode   options.StatsMode
-	txControl   *query.TransactionControl
-	syntax      options.Syntax
-	params      *params.Parameters
-	poolID      string
-	callOptions []grpc.CallOption
-}
-
-func (s testExecuteSettings) PoolID() string {
-	return s.poolID
-}
-
-func (s testExecuteSettings) RetryOpts() []retry.Option {
-	return nil
-}
-
-func (s testExecuteSettings) StatsCallback() func(stats stats.QueryStats) {
-	return nil
-}
-
-func (s testExecuteSettings) ExecMode() options.ExecMode {
-	return s.execMode
-}
-
-func (s testExecuteSettings) StatsMode() options.StatsMode {
-	return s.statsMode
-}
-
-func (s testExecuteSettings) TxControl() *query.TransactionControl {
-	return s.txControl
-}
-
-func (s testExecuteSettings) Syntax() options.Syntax {
-	return s.syntax
-}
-
-func (s testExecuteSettings) Params() *params.Parameters {
-	return s.params
-}
-
-func (s testExecuteSettings) CallOptions() []grpc.CallOption {
-	return s.callOptions
-}
-
-var _ executeSettings = testExecuteSettings{}
-
-type txMock func() *internal.Control
-
-func (tx txMock) txControl() *internal.Control {
-	if tx == nil {
-		return internal.NewControl(internal.WithTxID(""))
-	}
-
-	return tx()
-}
-
-func TestExecuteSettings(t *testing.T) {
-	for _, tt := range []struct {
-		name     string
-		tx       txMock
-		txOpts   []options.Execute
-		settings executeSettings
-	}{
-		{
-			name: "WithTxID",
-			tx: func() *internal.Control {
-				return internal.NewControl(internal.WithTxID("test"))
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExecute,
-				statsMode: options.StatsModeNone,
-				txControl: query.TxControl(query.WithTxID("test")),
-				syntax:    options.SyntaxYQL,
-			},
-		},
-		{
-			name: "WithStats",
-			txOpts: []options.Execute{
-				options.WithStatsMode(options.StatsModeFull, nil),
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExecute,
-				statsMode: options.StatsModeFull,
-				txControl: query.TxControl(query.WithTxID("")),
-				syntax:    options.SyntaxYQL,
-			},
-		},
-		{
-			name: "WithExecMode",
-			txOpts: []options.Execute{
-				options.WithExecMode(options.ExecModeExplain),
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExplain,
-				statsMode: options.StatsModeNone,
-				txControl: query.TxControl(query.WithTxID("")),
-				syntax:    options.SyntaxYQL,
-			},
-		},
-		{
-			name: "WithSyntax",
-			txOpts: []options.Execute{
-				options.WithSyntax(options.SyntaxPostgreSQL),
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExecute,
-				statsMode: options.StatsModeNone,
-				txControl: query.TxControl(query.WithTxID("")),
-				syntax:    options.SyntaxPostgreSQL,
-			},
-		},
-		{
-			name: "WithGrpcOptions",
-			txOpts: []options.Execute{
-				options.WithCallOptions(grpc.CallContentSubtype("test")),
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExecute,
-				statsMode: options.StatsModeNone,
-				txControl: query.TxControl(query.WithTxID("")),
-				syntax:    options.SyntaxYQL,
-				callOptions: []grpc.CallOption{
-					grpc.CallContentSubtype("test"),
-				},
-			},
-		},
-		{
-			name: "WithParams",
-			txOpts: []options.Execute{
-				options.WithParameters(
-					params.Builder{}.Param("$a").Text("A").Build(),
-				),
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExecute,
-				statsMode: options.StatsModeNone,
-				txControl: query.TxControl(query.WithTxID("")),
-				syntax:    options.SyntaxYQL,
-				params:    params.Builder{}.Param("$a").Text("A").Build(),
-			},
-		},
-		{
-			name: "WithCommitTx",
-			txOpts: []options.Execute{
-				options.WithCommit(),
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExecute,
-				statsMode: options.StatsModeNone,
-				txControl: query.TxControl(query.WithTxID(""), query.CommitTx()),
-				syntax:    options.SyntaxYQL,
-				params:    nil,
-			},
-		},
-		{
-			name: "WithPoolID",
-			txOpts: []options.Execute{
-				options.WithPoolID("test-pool-id"),
-			},
-			settings: testExecuteSettings{
-				execMode:  options.ExecModeExecute,
-				statsMode: options.StatsModeNone,
-				txControl: query.TxControl(query.WithTxID("")),
-				syntax:    options.SyntaxYQL,
-				poolID:    "test-pool-id",
-			},
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			a := allocator.New()
-			settings := options.ExecuteSettings(
-				append(
-					[]options.Execute{options.WithTxControl(tt.tx.txControl())},
-					tt.txOpts...,
-				)...,
-			)
-			require.Equal(t, tt.settings.Syntax(), settings.Syntax())
-			require.Equal(t, tt.settings.ExecMode(), settings.ExecMode())
-			require.Equal(t, tt.settings.StatsMode(), settings.StatsMode())
-			require.Equal(t, tt.settings.PoolID(), settings.PoolID())
-			require.Equal(t, tt.settings.TxControl().ToYDB(a).String(), settings.TxControl().ToYDB(a).String())
-			require.Equal(t, tt.settings.Params().ToYDB(a), settings.Params().ToYDB(a))
-			require.Equal(t, tt.settings.CallOptions(), settings.CallOptions())
-		})
-	}
 }
